fixes ui issues on dex v2
